safeReg/requirements.txt
asgiref==3.7.2
colorama==0.4.6
crispy-bootstrap5==2024.2
Django==5.0.2
django-auto-logout==0.5.1
django-axes==6.4.0
django-crispy-forms==2.1
django-environ==0.11.2
django-formtools==2.5.1
django-otp==1.3.0
django-phonenumber-field==7.3.0
django-recaptcha==4.0.0
django-two-factor-auth==1.16.0
phonenumbers==8.13.32
pypng==0.20220715.0
pytz==2024.1
qrcode==7.4.2
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
